pg_stat_walwrites statistics view
This view will provide the details of how many blocks,
writes and the time taken to write the WAL data to disk
by the backend, walwriter and other background process.
Based on this data, it is configure walwriter to reduce
the load on the backend processes, thus it should improve
the performance.

All the stats are collected and stored in a shared memory
strucuture protected by the WALWriteLock