pg_stat_wal_receiver to display connected host
With the support of multi host connection string support
in PostgreSQL, it is possible for the user to specify the
multi host connection string in recovery.conf to avoid
breakdown of streaming replication.

The pg_stat_wal_receiver view is enhanced to support display
of primary host information from which the replication is
streaming currently.